개인정보 변경 작업중

+++ src/main/java/com/green/muziuniv_be_user/application/account/AccountService.java
@@ -35,6 +35,7 @@ public class AccountService {
               .build();
    }
 
+   
 
 
    public String encodePassword(String rawPassword) {
@@ -65,7 +66,7 @@ public class AccountService {
    }
 
 
-   public PrivacyGetRes selectMyPrivacy(int loginId) {
+   public PrivacyGetRes selectMyPrivacy(long loginId) {
       return accountMapper.selectMyPrivacy(loginId);
    }
