+++ src/main/java/com/green/muziuniv_be_user/application/account/privacyandpwd/model/PrivacyGetRes.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PrivacyGetRes {
-    private int loginId;
+    private long loginId;
     private String userName;
     private String address;
     private String phone;
